feat: Rejig function handling to be a part of TSqlExpressionBUilder
Also removes need  for ExpressionList, which was an artifact of the grammar.
