+++ core/src/test/scala/com/databricks/labs/remorph/parsers/tsql/TSqlExpressionBuilderSpec.scala
@@ -1,9 +1,8 @@
 package com.databricks.labs.remorph.parsers.tsql
 
+import com.databricks.labs.remorph.parsers.{intermediate => ir}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-
-import com.databricks.labs.remorph.parsers.intermediate._
 
 class TSqlExpressionBuilderSpec extends AnyWordSpec with TSqlParserTestCommon with Matchers {
 
@@ -11,148 +10,273 @@ class TSqlExpressionBuilderSpec extends AnyWordSpec with TSqlParserTestCommon wi
 
   "TSqlExpressionBuilder" should {
     "translate literals" in {
-      example("null", _.expression(), Literal(nullType = Some(NullType())))
-      example("1", _.expression(), Literal(integer = Some(1)))
-      example("1.1", _.expression(), Literal(float = Some(1.1f)))
-      example("'foo'", _.expression(), Literal(string = Some("foo")))
+      example("null", _.expression(), ir.Literal(nullType = Some(ir.NullType())))
+      example("1", _.expression(), ir.Literal(integer = Some(1)))
+      example("1.1", _.expression(), ir.Literal(float = Some(1.1f)))
+      example("'foo'", _.expression(), ir.Literal(string = Some("foo")))
     }
     // TODO: note that scientific notation and decimals are not correctly handled if we copy SnowFlake
     "translate scientific notation" ignore {
-      example("1.1e2", _.expression(), Literal(integer = Some(110)))
-      example("1.1e-2", _.expression(), Literal(float = Some(0.011f)))
-      example("1e2", _.expression(), Literal(integer = Some(100)))
-      example("0.123456789", _.expression(), Literal(double = Some(0.123456789)))
-      example("0.123456789e-1234", _.expression(), Literal(decimal = Some(Decimal("0.123456789e-1234", None, None))))
+      example("1.1e2", _.expression(), ir.Literal(integer = Some(110)))
+      example("1.1e-2", _.expression(), ir.Literal(float = Some(0.011f)))
+      example("1e2", _.expression(), ir.Literal(integer = Some(100)))
+      example("0.123456789", _.expression(), ir.Literal(double = Some(0.123456789)))
+      example(
+        "0.123456789e-1234",
+        _.expression(),
+        ir.Literal(decimal = Some(ir.Decimal("0.123456789e-1234", None, None))))
     }
     "translate simple numeric binary expressions" in {
-      example("1 + 2", _.expression(), Add(Literal(integer = Some(1)), Literal(integer = Some(2))))
-      example("1 - 2", _.expression(), Subtract(Literal(integer = Some(1)), Literal(integer = Some(2))))
-      example("1 * 2", _.expression(), Multiply(Literal(integer = Some(1)), Literal(integer = Some(2))))
-      example("1 / 2", _.expression(), Divide(Literal(integer = Some(1)), Literal(integer = Some(2))))
-      example("1 % 2", _.expression(), Mod(Literal(integer = Some(1)), Literal(integer = Some(2))))
-      example("'A' || 'B'", _.expression(), Concat(Literal(string = Some("A")), Literal(string = Some("B"))))
-      example("4 ^ 2", _.expression(), BitwiseXor(Literal(integer = Some(4)), Literal(integer = Some(2))))
+      example("1 + 2", _.expression(), ir.Add(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))))
+      example("1 - 2", _.expression(), ir.Subtract(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))))
+      example("1 * 2", _.expression(), ir.Multiply(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))))
+      example("1 / 2", _.expression(), ir.Divide(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))))
+      example("1 % 2", _.expression(), ir.Mod(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))))
+      example("'A' || 'B'", _.expression(), ir.Concat(ir.Literal(string = Some("A")), ir.Literal(string = Some("B"))))
+      example("4 ^ 2", _.expression(), ir.BitwiseXor(ir.Literal(integer = Some(4)), ir.Literal(integer = Some(2))))
     }
     "translate complex binary expressions" in {
-      example("a + b * 2", _.expression(), Add(Column("a"), Multiply(Column("b"), Literal(integer = Some(2)))))
-      example("(a + b) * 2", _.expression(), Multiply(Add(Column("a"), Column("b")), Literal(integer = Some(2))))
-      example("a & b | c", _.expression(), BitwiseOr(BitwiseAnd(Column("a"), Column("b")), Column("c")))
-      example("(a & b) | c", _.expression(), BitwiseOr(BitwiseAnd(Column("a"), Column("b")), Column("c")))
+      example(
+        "a + b * 2",
+        _.expression(),
+        ir.Add(ir.Column("a"), ir.Multiply(ir.Column("b"), ir.Literal(integer = Some(2)))))
+      example(
+        "(a + b) * 2",
+        _.expression(),
+        ir.Multiply(ir.Add(ir.Column("a"), ir.Column("b")), ir.Literal(integer = Some(2))))
+      example("a & b | c", _.expression(), ir.BitwiseOr(ir.BitwiseAnd(ir.Column("a"), ir.Column("b")), ir.Column("c")))
+      example(
+        "(a & b) | c",
+        _.expression(),
+        ir.BitwiseOr(ir.BitwiseAnd(ir.Column("a"), ir.Column("b")), ir.Column("c")))
       example(
         "a % 3 + b * 2 - c / 5",
         _.expression(),
-        Subtract(
-          Add(Mod(Column("a"), Literal(integer = Some(3))), Multiply(Column("b"), Literal(integer = Some(2)))),
-          Divide(Column("c"), Literal(integer = Some(5)))))
+        ir.Subtract(
+          ir.Add(
+            ir.Mod(ir.Column("a"), ir.Literal(integer = Some(3))),
+            ir.Multiply(ir.Column("b"), ir.Literal(integer = Some(2)))),
+          ir.Divide(ir.Column("c"), ir.Literal(integer = Some(5)))))
       example(
         "(a % 3 + b) * 2 - c / 5",
         _.expression(),
-        Subtract(
-          Multiply(Add(Mod(Column("a"), Literal(integer = Some(3))), Column("b")), Literal(integer = Some(2))),
-          Divide(Column("c"), Literal(integer = Some(5)))))
-      example(query = "a || b || c", _.expression(), Concat(Concat(Column("a"), Column("b")), Column("c")))
+        ir.Subtract(
+          ir.Multiply(
+            ir.Add(ir.Mod(ir.Column("a"), ir.Literal(integer = Some(3))), ir.Column("b")),
+            ir.Literal(integer = Some(2))),
+          ir.Divide(ir.Column("c"), ir.Literal(integer = Some(5)))))
+      example(
+        query = "a || b || c",
+        _.expression(),
+        ir.Concat(ir.Concat(ir.Column("a"), ir.Column("b")), ir.Column("c")))
     }
     "correctly apply operator precedence and associativity" in {
       example(
         "1 + -++-2",
         _.expression(),
-        Add(Literal(integer = Some(1)), UMinus(UPlus(UPlus(UMinus(Literal(integer = Some(2))))))))
+        ir.Add(ir.Literal(integer = Some(1)), ir.UMinus(ir.UPlus(ir.UPlus(ir.UMinus(ir.Literal(integer = Some(2))))))))
       example(
         "1 + ~ 2 * 3",
         _.expression(),
-        Add(Literal(integer = Some(1)), Multiply(BitwiseNot(Literal(integer = Some(2))), Literal(integer = Some(3)))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Multiply(ir.BitwiseNot(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))))
       example(
         "1 + -2 * 3",
         _.expression(),
-        Add(Literal(integer = Some(1)), Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))))
       example(
         "1 + -2 * 3 + 7 & 66",
         _.expression(),
-        BitwiseAnd(
-          Add(
-            Add(Literal(integer = Some(1)), Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))),
-            Literal(integer = Some(7))),
-          Literal(integer = Some(66))))
+        ir.BitwiseAnd(
+          ir.Add(
+            ir.Add(
+              ir.Literal(integer = Some(1)),
+              ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))),
+            ir.Literal(integer = Some(7))),
+          ir.Literal(integer = Some(66))))
       example(
         "1 + -2 * 3 + 7 ^ 66",
         _.expression(),
-        BitwiseXor(
-          Add(
-            Add(Literal(integer = Some(1)), Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))),
-            Literal(integer = Some(7))),
-          Literal(integer = Some(66))))
+        ir.BitwiseXor(
+          ir.Add(
+            ir.Add(
+              ir.Literal(integer = Some(1)),
+              ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))),
+            ir.Literal(integer = Some(7))),
+          ir.Literal(integer = Some(66))))
       example(
         "1 + -2 * 3 + 7 | 66",
         _.expression(),
-        BitwiseOr(
-          Add(
-            Add(Literal(integer = Some(1)), Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))),
-            Literal(integer = Some(7))),
-          Literal(integer = Some(66))))
+        ir.BitwiseOr(
+          ir.Add(
+            ir.Add(
+              ir.Literal(integer = Some(1)),
+              ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))),
+            ir.Literal(integer = Some(7))),
+          ir.Literal(integer = Some(66))))
       example(
         "1 + -2 * 3 + 7 + ~66",
         _.expression(),
-        Add(
-          Add(
-            Add(Literal(integer = Some(1)), Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))),
-            Literal(integer = Some(7))),
-          BitwiseNot(Literal(integer = Some(66)))))
+        ir.Add(
+          ir.Add(
+            ir.Add(
+              ir.Literal(integer = Some(1)),
+              ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))),
+            ir.Literal(integer = Some(7))),
+          ir.BitwiseNot(ir.Literal(integer = Some(66)))))
       example(
         "1 + -2 * 3 + 7 | 1980 || 'leeds1' || 'leeds2' || 'leeds3'",
         _.expression(),
-        Concat(
-          Concat(
-            Concat(
-              BitwiseOr(
-                Add(
-                  Add(
-                    Literal(integer = Some(1)),
-                    Multiply(UMinus(Literal(integer = Some(2))), Literal(integer = Some(3)))),
-                  Literal(integer = Some(7))),
-                Literal(integer = Some(1980))),
-              Literal(string = Some("leeds1"))),
-            Literal(string = Some("leeds2"))),
-          Literal(string = Some("leeds3"))))
+        ir.Concat(
+          ir.Concat(
+            ir.Concat(
+              ir.BitwiseOr(
+                ir.Add(
+                  ir.Add(
+                    ir.Literal(integer = Some(1)),
+                    ir.Multiply(ir.UMinus(ir.Literal(integer = Some(2))), ir.Literal(integer = Some(3)))),
+                  ir.Literal(integer = Some(7))),
+                ir.Literal(integer = Some(1980))),
+              ir.Literal(string = Some("leeds1"))),
+            ir.Literal(string = Some("leeds2"))),
+          ir.Literal(string = Some("leeds3"))))
     }
     "correctly respect explicit precedence with parentheses" in {
       example(
         "(1 + 2) * 3",
         _.expression(),
-        Multiply(Add(Literal(integer = Some(1)), Literal(integer = Some(2))), Literal(integer = Some(3))))
+        ir.Multiply(
+          ir.Add(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))),
+          ir.Literal(integer = Some(3))))
       example(
         "1 + (2 * 3)",
         _.expression(),
-        Add(Literal(integer = Some(1)), Multiply(Literal(integer = Some(2)), Literal(integer = Some(3)))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Multiply(ir.Literal(integer = Some(2)), ir.Literal(integer = Some(3)))))
       example(
         "(1 + 2) * (3 + 4)",
         _.expression(),
-        Multiply(
-          Add(Literal(integer = Some(1)), Literal(integer = Some(2))),
-          Add(Literal(integer = Some(3)), Literal(integer = Some(4)))))
+        ir.Multiply(
+          ir.Add(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(2))),
+          ir.Add(ir.Literal(integer = Some(3)), ir.Literal(integer = Some(4)))))
       example(
         "1 + (2 * 3) + 4",
         _.expression(),
-        Add(
-          Add(Literal(integer = Some(1)), Multiply(Literal(integer = Some(2)), Literal(integer = Some(3)))),
-          Literal(integer = Some(4))))
+        ir.Add(
+          ir.Add(
+            ir.Literal(integer = Some(1)),
+            ir.Multiply(ir.Literal(integer = Some(2)), ir.Literal(integer = Some(3)))),
+          ir.Literal(integer = Some(4))))
       example(
         "1 + (2 * 3 + 4)",
         _.expression(),
-        Add(
-          Literal(integer = Some(1)),
-          Add(Multiply(Literal(integer = Some(2)), Literal(integer = Some(3))), Literal(integer = Some(4)))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Add(
+            ir.Multiply(ir.Literal(integer = Some(2)), ir.Literal(integer = Some(3))),
+            ir.Literal(integer = Some(4)))))
       example(
         "1 + (2 * (3 + 4))",
         _.expression(),
-        Add(
-          Literal(integer = Some(1)),
-          Multiply(Literal(integer = Some(2)), Add(Literal(integer = Some(3)), Literal(integer = Some(4))))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Multiply(
+            ir.Literal(integer = Some(2)),
+            ir.Add(ir.Literal(integer = Some(3)), ir.Literal(integer = Some(4))))))
       example(
         "(1 + (2 * (3 + 4)))",
         _.expression(),
-        Add(
-          Literal(integer = Some(1)),
-          Multiply(Literal(integer = Some(2)), Add(Literal(integer = Some(3)), Literal(integer = Some(4))))))
+        ir.Add(
+          ir.Literal(integer = Some(1)),
+          ir.Multiply(
+            ir.Literal(integer = Some(2)),
+            ir.Add(ir.Literal(integer = Some(3)), ir.Literal(integer = Some(4))))))
+    }
+    "translate functions with no parameters" in {
+      example("APP_NAME()", _.expression(), ir.CallFunction("APP_NAME", List()))
+      example("SCOPE_IDENTITY()", _.expression(), ir.CallFunction("SCOPE_IDENTITY", List()))
+    }
+
+    "translate functions with variable numbers of parameters" in {
+      example(
+        "CONCAT('a', 'b', 'c')",
+        _.expression(),
+        ir.CallFunction(
+          "CONCAT",
+          Seq(ir.Literal(string = Some("a")), ir.Literal(string = Some("b")), ir.Literal(string = Some("c")))))
+
+      example(
+        "CONCAT_WS(',', 'a', 'b', 'c')",
+        _.expression(),
+        ir.CallFunction(
+          "CONCAT_WS",
+          List(
+            ir.Literal(string = Some(",")),
+            ir.Literal(string = Some("a")),
+            ir.Literal(string = Some("b")),
+            ir.Literal(string = Some("c")))))
+    }
+
+    "translate functions with functions as parameters" in {
+      example(
+        "CONCAT(Greatest(42, 2, 4, \"ali\"), 'c')",
+        _.expression(),
+        ir.CallFunction(
+          "CONCAT",
+          List(
+            ir.CallFunction(
+              "Greatest",
+              List(
+                ir.Literal(integer = Some(42)),
+                ir.Literal(integer = Some(2)),
+                ir.Literal(integer = Some(4)),
+                ir.Column("\"ali\""))),
+            ir.Literal(string = Some("c")))))
+    }
+
+    "translate functions with complicated expressions as parameters" in {
+      example(
+        "CONCAT('a', 'b' || 'c', Greatest(42, 2, 4, \"ali\"))",
+        _.standardFunction(),
+        ir.CallFunction(
+          "CONCAT",
+          List(
+            ir.Literal(string = Some("a")),
+            ir.Concat(ir.Literal(string = Some("b")), ir.Literal(string = Some("c"))),
+            ir.CallFunction(
+              "Greatest",
+              List(
+                ir.Literal(integer = Some(42)),
+                ir.Literal(integer = Some(2)),
+                ir.Literal(integer = Some(4)),
+                ir.Column("\"ali\""))))))
+    }
+
+    "translate unknown functions as unresolved" in {
+      example(
+        "UNKNOWN_FUNCTION()",
+        _.expression(),
+        ir.UnresolvedFunction("UNKNOWN_FUNCTION", List(), is_distinct = false, is_user_defined_function = false))
+    }
+
+    "throw an exception for invalid function arguments" in {
+      // Later, we will not throw an exception but register a semantic error
+      intercept[IllegalArgumentException] {
+        example(
+          "USER_NAME('a', 'b', 'c', 'd')", // USER_NAME function only accepts 0 or 1 argument
+          _.expression(),
+          ir.Noop)
+      }
+      intercept[IllegalArgumentException] {
+        example(
+          "FLOOR()", // FLOOR requires 1 argument
+          _.expression(),
+          ir.Noop)
+      }
     }
   }
 }
